Did some improvements to the notifications.
Now it load notifications from db everytime the user navigate in the app IF THERE IS A PROBLEM WITH WEB SOCKET SERVER.
Visiting the home page syncronizes the notifications with the db even if the web socket server is working.
ws_token is saved to the session storage in the browser and it is unset from the php session when the user first visit the home page after log in.